fixed issue with loading hcl or die
